+++ src/main/java/kitchenpos/menu/application/MenuService.java
@@ -1,46 +1,33 @@
 package kitchenpos.menu.application;
 
-import kitchenpos.menu.repository.MenuGroupRepository;
-import kitchenpos.menu.repository.MenuRepository;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.dto.MenuRequest;
 import kitchenpos.menu.dto.MenuResponse;
+import kitchenpos.menu.repository.MenuRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
-import static java.util.stream.Collectors.groupingBy;
-
 @Transactional(readOnly = true)
 @Service
 public class MenuService {
     private final MenuRepository menuRepository;
-    private final MenuGroupRepository menuGroupRepository;
     private final MenuValidator menuValidator;
 
     public MenuService(
             final MenuRepository menuRepository,
-            final MenuGroupRepository menuGroupRepository,
             final MenuValidator menuValidator
     ) {
         this.menuRepository = menuRepository;
-        this.menuGroupRepository = menuGroupRepository;
         this.menuValidator = menuValidator;
     }
 
     @Transactional
     public MenuResponse create(MenuRequest request) {
-        validateMenuGroup(request.getMenuGroupId());
         menuValidator.validate(request);
         final Menu menu = menuRepository.save(request.toMenu());
         return MenuResponse.of(menu);
-    }
-
-    private void validateMenuGroup(Long menuGroupId) {
-        if (!menuGroupRepository.existsById(menuGroupId)) {
-            throw new IllegalArgumentException("존재하지 않는 메뉴 그룹입니다.");
-        }
     }
 
     public List<MenuResponse> list() {
