+++ src/main/java/kitchenpos/order/domain/Order.java
@@ -1,10 +1,7 @@
 package kitchenpos.order.domain;
 
-import kitchenpos.common.domain.OrderStatus;
-
 import javax.persistence.*;
 import java.time.LocalDateTime;
-import java.util.ArrayList;
 import java.util.List;
 
 import static java.util.Objects.requireNonNull;
